Refactored ES API queries
