+++ lib/account.js
@@ -3,7 +3,7 @@
 const logger = require('./logger');
 const Boom = require('@hapi/boom');
 const msgpack = require('msgpack5')();
-const { normalizePath, formatAccountListingResponse, mergeObjects } = require('./tools');
+const { normalizePath, formatAccountListingResponse, unpackUIDRangeForSearch, mergeObjects } = require('./tools');
 const crypto = require('crypto');
 const { MessageChannel } = require('worker_threads');
 const { MessagePortReadable } = require('./message-port-stream');
@@ -24,6 +24,8 @@ class Account {
         this.documentsQueue = options.documentsQueue || false;
 
         this.secret = options.secret;
+
+        this.esClient = options.esClient;
 
         this.call = options.call; // async method to request data from parent
 
@@ -598,7 +600,48 @@ class Account {
     }
 
     async getText(text, options) {
-        await this.loadAccountData(this.account, true);
+        await this.loadAccountData(this.account, !options.documentStore);
+
+        if (options.documentStore && (await settings.get('documentStoreEnabled'))) {
+            const { index, client } = this.esClient;
+
+            let buf = Buffer.from(text, 'base64url');
+            let message = buf.subarray(0, 8).toString('base64url');
+
+            let getResult = await client.get({
+                index,
+                id: `${this.account}:${message}`
+            });
+
+            this.logger.trace({
+                msg: 'Executed ES query',
+                query: { type: 'get', index, id: `${this.account}:${message}` },
+                results: getResult && getResult._source ? 1 : 0
+            });
+
+            if (!getResult || !getResult._source) {
+                let message = 'Requested message was not found';
+                let error = Boom.boomify(new Error(message), { statusCode: 404 });
+                throw error;
+            }
+
+            let messageData = getResult._source;
+            let response = {};
+
+            response.hasMore = false;
+            for (let textType of Object.keys(messageData.text || {})) {
+                if (['plain', 'html'].includes(textType) && (options.textType === '*' || options.textType === textType)) {
+                    if (options.maxBytes && messageData.text[textType].length > options.maxBytes) {
+                        response[textType] = messageData.text[textType].substring(0, options.maxBytes);
+                        response.hasMore = true;
+                    } else {
+                        response[textType] = messageData.text[textType];
+                    }
+                }
+            }
+
+            return response;
+        }
 
         return await this.call({ cmd: 'getText', account: this.account, text, options });
     }
@@ -606,8 +649,75 @@ class Account {
     async getMessage(message, options) {
         await this.loadAccountData(this.account, !options.documentStore);
 
-        if (options.documentStore) {
-            // fetch from cache instead
+        if (options.documentStore && (await settings.get('documentStoreEnabled'))) {
+            const { index, client } = this.esClient;
+
+            const reqOpts = {
+                index,
+                id: `${this.account}:${message}`,
+                _source_excludes: 'preview,seemsLikeNew'
+            };
+
+            switch (options.textType) {
+                case '*':
+                    break;
+                case 'html':
+                    reqOpts._source_excludes = 'text.plain';
+                    break;
+                case 'plain':
+                    reqOpts._source_excludes = 'text.html';
+                    break;
+                default:
+                    reqOpts._source_excludes = 'text.plain,text.html';
+            }
+
+            let getResult = await client.get(reqOpts);
+
+            this.logger.trace({
+                msg: 'Executed ES query',
+                query: { type: 'get', index, id: `${this.account}:${message}` },
+                results: getResult && getResult._source ? 1 : 0
+            });
+
+            if (!getResult || !getResult._source) {
+                let message = 'Requested message was not found';
+                let error = Boom.boomify(new Error(message), { statusCode: 404 });
+                throw error;
+            }
+
+            let messageData = getResult._source;
+
+            // restore headers and text object as per the API response
+            let headersObj = {};
+            for (let { key, value } of messageData.headers) {
+                headersObj[key] = value;
+            }
+            messageData.headers = headersObj;
+
+            if (messageData.text && (messageData.text.html || messageData.text.plain)) {
+                messageData.text.hasMore = false;
+
+                for (let textType of Object.keys(messageData.text || {})) {
+                    if (['plain', 'html'].includes(textType) && options.maxBytes && messageData.text[textType].length > options.maxBytes) {
+                        messageData.text[textType] = messageData.text[textType].substring(0, options.maxBytes);
+                        messageData.text.hasMore = true;
+                    }
+                }
+            }
+
+            for (let key of ['unseen', 'flagged', 'answered', 'draft']) {
+                if (messageData[key] === false) {
+                    messageData[key] = undefined;
+                }
+            }
+
+            for (let key of ['account', 'created', 'specialUse']) {
+                if (messageData[key]) {
+                    messageData[key] = undefined;
+                }
+            }
+
+            return messageData;
         }
 
         let messageData = await this.call({ cmd: 'getMessage', account: this.account, message, options });
@@ -621,6 +731,436 @@ class Account {
 
     async listMessages(query) {
         await this.loadAccountData(this.account, true);
+
+        if (query.documentStore && (await settings.get('documentStoreEnabled'))) {
+            const { index, client } = this.esClient;
+
+            let inboxData = await this.redis.hgetBuffer(this.getMailboxListKey(), 'INBOX');
+            let delimiter;
+            if (inboxData) {
+                try {
+                    inboxData = msgpack.decode(inboxData);
+                    delimiter = inboxData.delimiter;
+                } catch (err) {
+                    delimiter = '/'; // hope for the best
+                    inboxData = false;
+                }
+            }
+
+            inboxData = inboxData || {
+                path: 'INBOX',
+                delimiter
+            };
+
+            inboxData.specialUse = inboxData.specialUse || '\\Inbox';
+
+            let path = normalizePath(query.path, delimiter);
+            let mailboxData = path === 'INBOX' ? inboxData : false;
+            if (!mailboxData) {
+                mailboxData = await this.redis.hgetBuffer(this.getMailboxListKey(), path);
+                if (mailboxData) {
+                    try {
+                        mailboxData = msgpack.decode(mailboxData);
+                    } catch (err) {
+                        mailboxData = false;
+                    }
+                }
+            }
+
+            let searchQuery = {
+                bool: {
+                    must: [
+                        {
+                            term: {
+                                account: this.account
+                            }
+                        }
+                    ]
+                }
+            };
+
+            searchQuery.bool.must.push({
+                bool: {
+                    should: [
+                        {
+                            term: {
+                                path
+                            }
+                        },
+                        {
+                            term: {
+                                labels: mailboxData.specialUse || path
+                            }
+                        }
+                    ],
+                    minimum_should_match: 1
+                }
+            });
+
+            let searchResult = await client.search({
+                index,
+                size: query.pageSize,
+                from: query.pageSize * query.page,
+                query: searchQuery,
+                sort: { uid: 'desc' },
+                _source_excludes: 'headers,text.plain,text.html,seemsLikeNew'
+            });
+
+            this.logger.trace({
+                msg: 'Executed ES query',
+                query: { type: 'search', index, size: query.pageSize, from: query.pageSize * query.page, query: searchQuery, sort: { uid: 'desc' } },
+                results: searchResult.hits.total.value
+            });
+
+            let response = {
+                total: searchResult.hits.total.value,
+                page: query.page,
+                pages: Math.max(Math.ceil(searchResult.hits.total.value / query.pageSize), 1),
+                messages: searchResult.hits.hits.map(entry => {
+                    let messageData = entry._source;
+
+                    // normalize as per the API response
+
+                    for (let key of ['unseen', 'flagged', 'answered', 'draft']) {
+                        if (messageData[key] === false) {
+                            messageData[key] = undefined;
+                        }
+                    }
+
+                    for (let key of ['account', 'created', 'specialUse']) {
+                        if (messageData[key]) {
+                            messageData[key] = undefined;
+                        }
+                    }
+
+                    return messageData;
+                })
+            };
+
+            return response;
+        }
+
+        let path = normalizePath(query.path);
+        let encodedMailboxData = await this.redis.hgetBuffer(this.getMailboxListKey(), path);
+        if (!encodedMailboxData) {
+            let message = 'Mailbox record was not found';
+            let error = Boom.boomify(new Error(message), { statusCode: 404 });
+            error.output.payload.path = query.path;
+            throw error;
+        }
+
+        // mailbox seems to exist, so call parent to resolve open connection
+        return await this.call(Object.assign({ cmd: 'listMessages', account: this.account }, query));
+    }
+
+    async searchMessages(query) {
+        await this.loadAccountData(this.account, true);
+
+        if (query.documentStore && (await settings.get('documentStoreEnabled'))) {
+            const { index, client } = this.esClient;
+
+            let searchQuery = {
+                bool: {
+                    must: [
+                        {
+                            term: {
+                                account: this.account
+                            }
+                        }
+                    ]
+                }
+            };
+
+            if (query.path) {
+                let inboxData = await this.redis.hgetBuffer(this.getMailboxListKey(), 'INBOX');
+                let delimiter;
+
+                if (inboxData) {
+                    try {
+                        inboxData = msgpack.decode(inboxData);
+                        delimiter = inboxData.delimiter;
+                    } catch (err) {
+                        delimiter = '/'; // hope for the best
+                        inboxData = false;
+                    }
+                }
+
+                inboxData = inboxData || {
+                    path: 'INBOX',
+                    delimiter
+                };
+
+                inboxData.specialUse = inboxData.specialUse || '\\Inbox';
+
+                let path = normalizePath(query.path, delimiter);
+                let mailboxData = path === 'INBOX' ? inboxData : false;
+                if (!mailboxData) {
+                    mailboxData = await this.redis.hgetBuffer(this.getMailboxListKey(), path);
+                    if (mailboxData) {
+                        try {
+                            mailboxData = msgpack.decode(mailboxData);
+                        } catch (err) {
+                            mailboxData = false;
+                        }
+                    }
+                }
+
+                searchQuery.bool.must.push({
+                    bool: {
+                        should: [
+                            {
+                                term: {
+                                    path
+                                }
+                            },
+                            {
+                                term: {
+                                    labels: mailboxData.specialUse || path
+                                }
+                            }
+                        ],
+                        minimum_should_match: 1
+                    }
+                });
+            }
+
+            for (let key of ['answered', 'deleted', 'draft', 'unseen', 'flagged']) {
+                if (typeof query.search[key] === 'boolean') {
+                    searchQuery.bool.must.push({
+                        term: {
+                            [key]: query.search[key]
+                        }
+                    });
+                }
+            }
+
+            if (typeof query.search.seen === 'boolean') {
+                searchQuery.bool.must.push({
+                    term: {
+                        unseen: !query.search.seen
+                    }
+                });
+            }
+
+            for (let key of ['from', 'to', 'cc', 'bcc']) {
+                if (query.search[key]) {
+                    searchQuery.bool.must.push({
+                        bool: {
+                            should: [
+                                {
+                                    match: {
+                                        [`${key}.name`]: {
+                                            query: query.search[key],
+                                            operator: 'and'
+                                        }
+                                    }
+                                },
+                                {
+                                    term: {
+                                        [`${key}.address`]: query.search[key]
+                                    }
+                                }
+                            ],
+                            minimum_should_match: 1
+                        }
+                    });
+                }
+            }
+
+            if (query.search.uid) {
+                let uidEntries = unpackUIDRangeForSearch(query.search.uid);
+                if (uidEntries && uidEntries.length) {
+                    let mustList = [];
+                    for (let entry of uidEntries) {
+                        if (typeof entry === 'number') {
+                            mustList.push({
+                                match: {
+                                    uid: {
+                                        query: entry,
+                                        operator: 'and'
+                                    }
+                                }
+                            });
+                        } else if (typeof entry === 'object') {
+                            mustList.push({
+                                range: {
+                                    uid: entry
+                                }
+                            });
+                        }
+                    }
+
+                    if (mustList.length) {
+                        searchQuery.bool.must.push({
+                            bool: {
+                                should: mustList,
+                                minimum_should_match: 1
+                            }
+                        });
+                    }
+                }
+            }
+
+            for (let key of ['emailId', 'threadId']) {
+                if (query.search[key]) {
+                    searchQuery.bool.must.push({
+                        term: {
+                            key: query.search[key]
+                        }
+                    });
+                }
+            }
+
+            if (query.search.subject) {
+                searchQuery.bool.must.push({
+                    match: {
+                        subject: {
+                            query: query.search.subject,
+                            operator: 'and'
+                        }
+                    }
+                });
+            }
+
+            if (query.search.body) {
+                searchQuery.bool.must.push({
+                    bool: {
+                        should: [
+                            {
+                                match: {
+                                    'text.plain': {
+                                        query: query.search.body,
+                                        operator: 'and'
+                                    }
+                                }
+                            },
+                            {
+                                match: {
+                                    'text.html': {
+                                        query: query.search.body,
+                                        operator: 'and'
+                                    }
+                                }
+                            }
+                        ],
+                        minimum_should_match: 1
+                    }
+                });
+            }
+
+            let dateMatch = {};
+
+            for (let key of ['before', 'sentBefore']) {
+                if (query.search[key]) {
+                    dateMatch.lte = query.search[key];
+                }
+            }
+
+            for (let key of ['since', 'sentSince']) {
+                if (query.search[key]) {
+                    dateMatch.gte = query.search[key];
+                }
+            }
+
+            if (Object.keys(dateMatch).length) {
+                searchQuery.bool.must.push({
+                    range: { date: dateMatch }
+                });
+            }
+
+            let sizeMatch = {};
+
+            if (query.search.larger) {
+                dateMatch.gte = query.search.larger;
+            }
+
+            if (query.search.smaller) {
+                dateMatch.lte = query.search.smaller;
+            }
+
+            if (Object.keys(sizeMatch).length) {
+                searchQuery.bool.must.push({
+                    range: { size: sizeMatch }
+                });
+            }
+
+            // headers, nested query
+            if (Object.keys(query.search.header || {}).length) {
+                Object.keys(query.search.header).forEach(header => {
+                    searchQuery.bool.must.push({
+                        nested: {
+                            path: 'headers',
+                            query: {
+                                bool: {
+                                    must: [
+                                        {
+                                            term: {
+                                                'headers.key': header.toLowerCase()
+                                            }
+                                        },
+                                        {
+                                            match: {
+                                                'headers.value': {
+                                                    query: (query.search.header[header] || '').toString(),
+                                                    operator: 'and'
+                                                }
+                                            }
+                                        }
+                                    ]
+                                }
+                            }
+                        }
+                    });
+                });
+            }
+
+            let searchResult = await client.search({
+                index,
+                size: query.pageSize,
+                from: query.pageSize * query.page,
+                query: searchQuery,
+                sort: { uid: 'desc' },
+                _source_excludes: 'headers,text.plain,text.html,seemsLikeNew'
+            });
+
+            this.logger.trace({
+                msg: 'Executed ES query',
+                query: { type: 'search', index, size: query.pageSize, from: query.pageSize * query.page, query: searchQuery, sort: { uid: 'desc' } },
+                results: searchResult.hits.total.value
+            });
+
+            let response = {
+                total: searchResult.hits.total.value,
+                page: query.page,
+                pages: Math.max(Math.ceil(searchResult.hits.total.value / query.pageSize), 1)
+            };
+
+            if (query.exposeQuery) {
+                response.documentStoreQuery = query;
+            }
+
+            response.messages = searchResult.hits.hits.map(entry => {
+                let messageData = entry._source;
+
+                // normalize as per the API response
+
+                for (let key of ['unseen', 'flagged', 'answered', 'draft']) {
+                    if (messageData[key] === false) {
+                        messageData[key] = undefined;
+                    }
+                }
+
+                for (let key of ['account', 'created', 'specialUse']) {
+                    if (messageData[key]) {
+                        messageData[key] = undefined;
+                    }
+                }
+
+                return messageData;
+            });
+
+            return response;
+        }
 
         let path = normalizePath(query.path);
         let encodedMailboxData = await this.redis.hgetBuffer(this.getMailboxListKey(), path);
