+++ lib/tools.js
@@ -662,22 +662,28 @@ MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEV3QUiYsp13nD9suD1/ZkEXnuMoSg
     },
 
     async failAction(request, h, err) {
-        let details = (err.details || []).map(detail => ({ message: detail.message, key: detail.context.key }));
+        try {
+            let details = (err.details || []).map(detail => ({ message: detail.message, key: detail.context.key }));
 
-        delete err._original;
+            delete err._original;
 
-        logger.error({
-            msg: 'Request failed',
-            method: request.method,
-            route: request.route.path,
-            statusCode: request.response && request.response.statusCode,
-            err
-        });
+            logger.error({
+                msg: 'Request failed',
+                method: request.method,
+                route: request.route.path,
+                statusCode: request.response && request.response.statusCode,
+                err
+            });
 
-        let message = 'Invalid input';
-        let error = Boom.boomify(new Error(message), { statusCode: 400 });
-        error.output.payload.fields = details;
-        throw error;
+            let message = 'Invalid input';
+            let error = Boom.boomify(new Error(message), { statusCode: 400 });
+            error.output.payload.fields = details;
+
+            throw error;
+        } catch (E) {
+            request.logger.error({ err: E });
+            throw E;
+        }
     },
 
     async assertPreconditions(redis) {
